chore: Update application.yaml (#51)
Updating application.yaml according to openshift.io use case

Changes did -

1. Adds parameter - RELEASE_VERSION which is basically the
build number and will be used to differentiate between
the resources of the particular build.

2. Adds parameter - SUFFIX_NAME which is an identifier to
differentiate between the master build or other branches.
More like a master build or PR build. So whenever there is
a change in the resources of a particular branch, it applies
to the resource of that branch only. The SUFFIX_NAME variable
is used in the name of all resources.

3. Removes image stream for s2i base Docker image because it
will trigger a build whenever change that is not our use case
and also may be breaking sometime for the user or while
processing template it is creating two builds.

4. Adds a field label in image stream to tag the images
generated with the RELEASE_VERSION.

5. Removes  trigger from build config because
we kept at one place in deployment Config. Keeping at both
places are triggering two deployments.